`destroy` asks for confirmation. [GH-699]

+++ lib/vagrant/command/destroy.rb
@@ -15,8 +15,17 @@ module Vagrant
         @logger.debug("'Destroy' each target VM...")
         with_target_vms(argv[0]) do |vm|
           if vm.created?
-            @logger.info("Destroying: #{vm.name}")
-            vm.destroy
+            choice = @env.ui.ask(I18n.t("vagrant.commands.destroy.confirmation",
+                                        :name => vm.name))
+
+            if choice.upcase == "Y"
+              @logger.info("Destroying: #{vm.name}")
+              vm.destroy
+            else
+              @logger.info("Not destroying #{vm.name} since confirmation was declined.")
+              @env.ui.success(I18n.t("vagrant.commands.destroy.will_not_destroy",
+                                     :name => vm.name), :prefix => false)
+            end
           else
             @logger.info("Not destroying #{vm.name}, since it isn't created.")
             vm.ui.info I18n.t("vagrant.commands.common.vm_not_created")
